Require a subpackage of ginkgo
This is a hack.

We depend on the vendored packages in our CI system, which is a good
approach when we vendor all dependencies in the repo...

Moving to go modules, means we're vendoring only things that we're
actually using, therefore some dependencies are slimmed out. We can
carry on with our patter by making our tests require ginkgo CLI for
running but not use it after all.

That will trick the go modules into downloading the required files that
are then used for `go install`